Prevents errors due to not enough attributes
Since my method of array population was to create an array of numbers, and
slowly fill it in with data, checking to see if the array was full of numbers is
a way to see if all the attributes are collected. This commit implements said
check, and only writes if all of the attributes are collected.

This prevents errors converting `int`s to `str`s.